LayoutTests/inspector/model/remote-object/string-large.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/remote-object-utilities.js"></script>
<script>
function test()
{
    let steps = [
        {
            expression: `"0123456789".repeat(1000000 / 10)`,
            filter(key, value) {
                if (key === "_description" || key === "_value") {
                    InspectorTest.assert(value === "0123456789".repeat(1000000 / 10));
                    return true;
                }
                return false;
            },
        },
    ];

    if (!window.WI) {
        window.steps = steps;
        return;
    }

    runSteps(steps);
}
</script>
</head>
<body onload="runTest(); runInBrowserTest();"></body>
</html>
