LayoutTests/inspector/model/remote-object/string-large-expected.txt


-----------------------------------------------------
EXPRESSION: "0123456789".repeat(1000000 / 10)
{
  "_type": "string",
  "_description": "<filtered>",
  "_value": "<filtered>"
}
